Cleanup test reuse
